SolnAI-agents/example-agent/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
supabase==2.12.0
python-dotenv==1.0.1
requests==2.25.1
httpx==0.28.1
openai==0.28.1
pytest==6.2.5
pytest-asyncio==0.21.1
